+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ### 🧰 Bug fixes 🧰
 - Fix translation from otlp.Request to pdata representation, changes to the returned pdata not all reflected to the otlp.Request (#5197)
+- `exporterhelper` now properly consumes any remaining items on stop (#5203)
 
 ## v0.49.0 Beta